Changes to define and use a timestamp for use in identifying a log of cmd output and a folder to store key output files.
LogRunModel defines a timestamp to identify the run, and runs the model logging cmd output to a logfile named with the timestamp.  The same timestamp will be used to name the output folder in which key output files will be stored.  This commit includes related changes to SetPath.bat and SetUpModel.bat.